Edit of the index.html file to add some consistency
Creation of a meta title tag.
Addition of content in the description meta tag.
Addition of the keyword in the h2 tag.
Edit of 6 alts from img tag.
Replacement of useless images with texts.
Replacement of the "page2>" link to "contact".
Replacement of div elements to sections and articles to ass consistency.